+++ components/camel-google/camel-google-pubsub/src/test/java/org/apache/camel/component/google/pubsub/integration/BodyTypesIT.java
@@ -16,6 +16,10 @@
  */
 package org.apache.camel.component.google.pubsub.integration;
 
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.io.ObjectInputStream;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -26,19 +30,23 @@ import org.apache.camel.Exchange;
 import org.apache.camel.Produce;
 import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.google.pubsub.GooglePubsubConstants;
 import org.apache.camel.component.google.pubsub.PubsubTestSupport;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.apache.camel.support.DefaultExchange;
 import org.junit.jupiter.api.Test;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertSame;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class SingleExchangeRoundtripTest extends PubsubTestSupport {
+public class BodyTypesIT extends PubsubTestSupport {
+    private static final Logger LOG = LoggerFactory.getLogger(BodyTypesIT.class);
 
-    private static final String TOPIC_NAME = "singleSend";
-    private static final String SUBSCRIPTION_NAME = "singleReceive";
+    private static final String TOPIC_NAME = "typesSend";
+    private static final String SUBSCRIPTION_NAME = "TypesReceive";
 
     @EndpointInject("direct:from")
     private Endpoint directIn;
@@ -60,11 +68,16 @@ public class SingleExchangeRoundtripTest extends PubsubTestSupport {
 
     @Override
     public void createTopicSubscription() {
-        createTopicSubscriptionPair(TOPIC_NAME, SUBSCRIPTION_NAME);
+        try {
+            createTopicSubscriptionPair(TOPIC_NAME, SUBSCRIPTION_NAME);
+        } catch (Exception e) {
+            // May be ignored because it could have been created. 
+            LOG.warn("Failed to create the subscription pair {}", e.getMessage());
+        }
     }
 
     @Override
-    protected RouteBuilder createRouteBuilder() {
+    protected RouteBuilder createRouteBuilder() throws Exception {
         return new RouteBuilder() {
             public void configure() {
                 from(directIn).routeId("Single_Send").to(pubsubTopic).to(sendResult);
@@ -77,21 +90,15 @@ public class SingleExchangeRoundtripTest extends PubsubTestSupport {
     }
 
     @Test
-    public void testSingleMessageSend() throws Exception {
+    public void byteArray() throws Exception {
 
         Exchange exchange = new DefaultExchange(context);
 
-        String attributeKey = "ATTRIBUTE-TEST-KEY";
-        String attributeValue = "ATTRIBUTE-TEST-VALUE";
+        byte[] body = { 1, 2, 3 };
 
-        Map<String, String> attributes = new HashMap<>();
-        attributes.put(attributeKey, attributeValue);
-
-        exchange.getIn().setBody("Single  : " + exchange.getExchangeId());
-        exchange.getIn().setHeader(GooglePubsubConstants.ATTRIBUTES, attributes);
+        exchange.getIn().setBody(body);
 
         receiveResult.expectedMessageCount(1);
-        receiveResult.expectedBodiesReceivedInAnyOrder(exchange.getIn().getBody());
 
         producer.send(exchange);
 
@@ -100,8 +107,9 @@ public class SingleExchangeRoundtripTest extends PubsubTestSupport {
 
         Exchange sentExchange = sentExchanges.get(0);
 
-        assertEquals(exchange.getIn().getHeader(GooglePubsubConstants.MESSAGE_ID),
-                sentExchange.getIn().getHeader(GooglePubsubConstants.MESSAGE_ID), "Sent ID");
+        assertTrue(sentExchange.getIn().getBody() instanceof byte[], "Sent body type is byte[]");
+
+        assertSame(body, sentExchange.getIn().getBody(), "Sent body type is the one sent");
 
         receiveResult.assertIsSatisfied(5000);
 
@@ -111,14 +119,52 @@ public class SingleExchangeRoundtripTest extends PubsubTestSupport {
 
         Exchange receivedExchange = receivedExchanges.get(0);
 
-        assertNotNull(receivedExchange.getIn().getHeader(GooglePubsubConstants.MESSAGE_ID), "PUBSUB Message ID Property");
-        assertNotNull(receivedExchange.getIn().getHeader(GooglePubsubConstants.PUBLISH_TIME), "PUBSUB Published Time");
+        assertTrue(receivedExchange.getIn().getBody() instanceof byte[], "Received body is of byte[] type");
 
-        assertEquals(attributeValue,
-                ((Map) receivedExchange.getIn().getHeader(GooglePubsubConstants.ATTRIBUTES)).get(attributeKey),
-                "PUBSUB Header Attribute");
+        assertTrue(Arrays.equals(body, (byte[]) receivedExchange.getIn().getBody()), "Received body equals sent");
 
-        assertEquals(sentExchange.getIn().getHeader(GooglePubsubConstants.MESSAGE_ID),
-                receivedExchange.getIn().getHeader(GooglePubsubConstants.MESSAGE_ID));
+    }
+
+    @Test
+    public void objectSerialised() throws Exception {
+
+        Exchange exchange = new DefaultExchange(context);
+
+        Map<String, String> body = new HashMap<>();
+        body.put("KEY", "VALUE1212");
+
+        exchange.getIn().setBody(body);
+
+        receiveResult.expectedMessageCount(1);
+
+        producer.send(exchange);
+
+        List<Exchange> sentExchanges = sendResult.getExchanges();
+        assertEquals(1, sentExchanges.size(), "Sent exchanges");
+
+        Exchange sentExchange = sentExchanges.get(0);
+
+        assertTrue(sentExchange.getIn().getBody() instanceof Map, "Sent body type is byte[]");
+
+        receiveResult.assertIsSatisfied(5000);
+
+        List<Exchange> receivedExchanges = receiveResult.getExchanges();
+
+        assertNotNull(receivedExchanges, "Received exchanges");
+
+        Exchange receivedExchange = receivedExchanges.get(0);
+
+        assertTrue(receivedExchange.getIn().getBody() instanceof byte[], "Received body is of byte[] type");
+
+        Object bodyReceived = deserialize((byte[]) receivedExchange.getIn().getBody());
+
+        assertEquals("VALUE1212", ((Map) bodyReceived).get("KEY"), "Received body is a Map");
+
+    }
+
+    public static Object deserialize(byte[] data) throws IOException, ClassNotFoundException {
+        ByteArrayInputStream in = new ByteArrayInputStream(data);
+        ObjectInputStream is = new ObjectInputStream(in);
+        return is.readObject();
     }
 }
